Fix Kotlin compile errors blocking the APK build.
Submit SMP actions as a Callable so StreamState is returned, and pass a Long to TimeUnit.toNanos.

+++ app/src/main/java/com/streaming/app/smp/SmpClient.kt
@@ -180,7 +180,7 @@ class SmpClient {
     private fun readResponse(inputStream: InputStream): String {
         val buffer = ByteArrayOutputStream()
         val chunk = ByteArray(1024)
-        val deadline = System.nanoTime() + TimeUnit.MILLISECONDS.toNanos(READ_TIMEOUT_MS)
+        val deadline = System.nanoTime() + TimeUnit.MILLISECONDS.toNanos(READ_TIMEOUT_MS.toLong())
 
         while (System.nanoTime() < deadline) {
             while (inputStream.available() > 0) {
